Fixed skull enabled animations
The burst animations when skulls are enabled at the start of every set are now correctly placed and scaled. The description text is kept in the same spot since it'll look like shit on the bottom. Also some preliminary work on a possible halo online weapon model swap later on.